Title: 시험 통과 여부 확인하기 2, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240211/시험 통과 여부 확인하기 2/verify-test-passed-2.cpp
@@ -6,7 +6,7 @@ int main() {
     int a;
     cin >> a;
 
-    int sum[10]; // 10명 미만의 학생들의 점수 저장
+    float sum[10]; // 10명 미만의 학생들의 점수 저장 (나중에 평균 계산해야 하니까 float)
 
     int cnt = 0; // 통과한 사람
